requirements.txt
## The following requirements were added by pip freeze:
anyascii==0.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==6.6
defusedxml==0.7.1
dj-database-url==0.4.1
Django==5.0.6
django-debug-toolbar==3.2.4
django-debugtools==2.0
django-filter==23.5
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
itsdangerous==1.1.0
l18n==2021.3
laces==0.1.1
MarkupSafe==1.1.1
openpyxl==3.1.4
pillow==10.3.0
pillow_heif==0.16.0
psycopg2-binary==2.9.9
python-decouple==3.6
python-dotenv==0.6.1
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.2
tablib==3.2.0
telepath==0.3.1
urllib3==2.2.2
wagtail==6.0
webencodings==0.5.1
Werkzeug==1.0.0
Willow==1.8.0
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
